backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Pillow
numpy
opencv-python-headless
torch
torchvision